+++ src/categories/categories.service.ts
@@ -6,7 +6,7 @@ import { PrismaService } from '../prisma.service';
 @Injectable()
 export class CategoriesService {
   constructor(private readonly prisma: PrismaService) {}
-  async create(data: CreateCategoryDto) {
+  create(data: CreateCategoryDto) {
     if (!data.name) throw new Error('Category name cannot be empty');
     return this.prisma.category.create({ data });
   }
@@ -14,12 +14,16 @@ export class CategoriesService {
     return this.prisma.category.findMany();
   }
   findOne(id: number) {
+    console.log('id::', id);
+    if (!id) throw new Error('Category ID is required');
     return this.prisma.category.findUnique({ where: { id } });
   }
   update(id: number, data: UpdateCategoryDto) {
+    if (!id || !data) throw new Error('error updating category');
     return this.prisma.category.update({ where: { id }, data });
   }
   remove(id: number) {
+    if (!id) throw new Error('Category ID is required');
     return this.prisma.category.delete({ where: { id } });
   }
 }
